Set content-type: ...; charset=utf-8 if a [non-empty] str is passed in as a Request's body.
(The reverse should probably also be done in Response.)

Make the content-type header's structure available as Request/Response.contenttype.

This is stepping into the realm of logic/behavior I don't want to manage/be responsible for 🙁. Ideally I'd use the Request/Response logic from urllib3, requests, etc. and only manage the low-level HTTP/2 I/O in nh2, but that's not obviously feasible.