Serialize all calls to Connection.flush using a lock (see #5).
I'm not sure why this is necessary. (anyio.SocketStream.send does do multiple `await`s, but I'm not sure if there is an actual potential for corruption, nor why anyio goes to the trouble of detecting concurrent sends but throws an error instead of using a lock itself.)

For now, additionally include all changes to the h2.connection.H2Connection object in the lock (so any time Connection.c is modified, the lock is held and remains held through the next Connection.flush()). This might be unnecessary, but, again, if all that's necessary is locking the call to self.s.send in Connection.flush, I don't know why anyio doesn't just do that itself 🙁.