Add a new LiveRequest.wait() (see #3).
When awaited, if nothing is currently running the parent Connection.read() in a loop, LiveRequest.wait() will do so, ensuring all incoming data is dispatched to the correct LiveRequest (eventually including itself).

If another task, R, is already running Connection.read() in a loop, the task awaiting await(), W, creates an Event and awaits it. R will wake W when either W's stream receives an h2.events.StreamEnded event (triggering a call to W's ended() method, which will set the event); or when R's own stream ends (causing it to pick another waiting task to "take over" as the new R).