Rename LiveRequest to Stream.
As part of #6, I had been thinking of renaming LiveRequest to LiveRex (to signify that it will represent both Requests and Responses), but there's a 1:1 relationship between in-protocol "streams" and live Requests (whenever a Request becomes a LiveRequest, it immediately becomes a numbered stream; the end of the stream signals the end of the LiveRequest being "live").

I do not like the confusion between an "HTTP/2 stream" and "TCP is a stream protocol" (as in SOCK_STREAM, anyio.EndOfStream, etc.). This is the standardized terminology, however 🙁.

A followup will probably remove Connection.new_stream.

It might also make sense to separate Stream.wait into explicit "wait for this live request to finish transmitting" and "wait for a response to this request".